Rename files with less ambiguous, generalized names
Jira ticket: CAMS-526
